GH-36: Move module and tag parsing to different files
